backend/requirements.txt
Django>=4.2
djangorestframework
python-dotenv
mysqlclient
gunicorn
django-cors-headers
psycopg2-binary

# 以下は任意（必要に応じて）
pytest
pytest-django
requests
ipython
Pillow
